Spawn a map, position the player, and add exploration system for debugging.
I use an exploration system to look around the map for my in-progress roguelike. Initial plans were to not use that in this jam, but I need to tweak the map and need exploration for debugging. So it, plus its integrations with other systems, have landed.

A 10X10 map of 10X10 rooms separated by exits is created. The player spawns at the bottom left, and a set of goal coordinates furthest from the player is chosen.

Ctrl+left/right/up/down arrow moves a virtual cursor around the map, announcing the content and direction/distance to the focused tile. Enter plots a path and navigates to the tile. Moving manually or via pathfinding automatically snaps exploration focus back to the player's location.